pyproject.toml
[tool.poetry]
name = "bot-jam.py"
version = "0.1.0"
description = ""
authors = ["timelessnesses <mooping3roblox@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = {git = "https://github.com/Rapptz/discord.py"}
orjson = "^3.6.7"
Flask = "^2.1.1"
aiohttp = {extras = ["speedups"], version = "^3.8.1"}
python-dotenv = "^0.20.0"
aiofiles = "^0.8.0"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
